Consolidate Chinese Zodiac UI from 9 files to 4
Merge YearSearch, YearDetail, and the AnimalInfo/ElementInfo/CombinedMeaning
cards into ChineseZodiac.jsx — they were leaf components used in one place,
each redefining identical Bullets/heading/chip helpers. Define those once and
add small BulletSection/ChipRow helpers to cut the repeated markup. YearTable
(the large grid), zodiacData, and zodiacUtils stay separate.
